+++ contributor-plugins/vcs/vcs-plugin-api/src/main/kotlin/com/mreil/easy/vcs/VcsGit.kt
@@ -39,7 +39,7 @@ internal class VcsGit(
 
     override fun push(): Provider<Boolean> = success("push")
 
-    override fun push(tag: String): Provider<Boolean> = success("push", "origin", "HEAD", tag)
+    override fun push(tag: String): Provider<Boolean> = success("push", "--atomic", "origin", "HEAD", tag)
 
     override fun fetch(): Provider<Boolean> = success("fetch")